fixed mock create/Delete dataset
Params are an array. For the sake of catching bugs I've added a generic
test for arrayness in mock Call, but this could be removed. Also to safely
deal with param arrays, and minimum args, I added getParamArray which tests
and generates the appropriate Error

Error codes were triggering lint when capitilized.