now retry-loop iscsi can't stat errors
iscsiadm seems to have some TOCTOU issues. When scanning the nodes, it may
find nodes that it wants to delete, and they they are deleted... this causes
failures, wehn if if it just retried, it would be fine.

We can't just skip at a higher layer as we may actually want a result