send targets before testing iscsi
when testing an iscsi setup when nothing is discovered, you need to send targets first
the issue then becomes that will error if there are no targets.

BUT I believe (although this is not 100% confirmed) that the discover without sendtargets
will then succeed, and eventually, when there are targets, then it will work.