fix iscsi locate --create not returning an error if the target is not discoverable
When the target is not discovered, we will attempt to create it. This is performed
via API, which means it may succeed, even when iSCSI itself cannot discover.

We later attempt to activate, but there are no targets still, even though
we created them, and thus we have nothing to activate, and thus no failure.

The solution is to test that we can find the newly created target after creating
it, if not that indicates and issue with discovery.

There is a situation where we may not discover all the created targets, but
if we have at least discovered one, then we can make progres, and in the incus
case we only ever perform a single activatation at a time.

When not using locate --create, we would not create the target, and thus we
would attempt to activate, and that would then fail correctly.